seapass_m/seapass-backend-main/seapass_m/seapass-backend-main/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
psycopg2-binary==2.9.11
python-dotenv==1.0.1
